src/run_nerf.py
# run_nerf.py
# Run NeRF on a dataset

import torch
import json
import argparse
import os
import numpy as np
from datetime import datetime
from tqdm import tqdm

device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
print(f"Torch version: {torch.__version__}, Torch Device: {device}")

# Define NeRF model architecture
class NeRFModel(torch.nn.Module):
    def __init__(self, D=8, W=256, input_ch=3, output_ch=4, skips=[4]):
        super(NeRFModel, self).__init__()
        self.D = D
        self.W = W
        self.input_ch = input_ch
        self.output_ch = output_ch
        self.skips = skips
        
        self.pts_linears = torch.nn.ModuleList(
            [torch.nn.Linear(input_ch, W)] + 
            [torch.nn.Linear(W, W) if i not in self.skips else torch.nn.Linear(W + input_ch, W) for i in range(D-1)]
        )
        self.output_linear = torch.nn.Linear(W, output_ch)
    
    def forward(self, x):
        h = x
        for i, l in enumerate(self.pts_linears):
            h = torch.relu(l(h))
            if i in self.skips:
                h = torch.cat([x, h], -1)
        outputs = self.output_linear(h)
        return outputs

# Data loader placeholder
def data_loader(data_path):
    # Implement data loading here
    # Return the dataset
    pass

# Training function
def train(model, dataloader, iterations, model_dir):
    optimizer = torch.optim.Adam(model.parameters(), lr=5e-4)
    criterion = torch.nn.MSELoss()
    model.train()
    
    for i in tqdm(range(iterations)):
        for data in dataloader:
            inputs, targets = data
            inputs, targets = inputs.to(device), targets.to(device)
            
            optimizer.zero_grad()
            outputs = model(inputs)
            loss = criterion(outputs, targets)
            loss.backward()
            optimizer.step()
        
        if (i + 1) % 1000 == 0:
            torch.save(model.state_dict(), os.path.join(model_dir, f"model_{i+1}.pth"))

# Testing function
def test(model, dataloader):
    model.eval()
    with torch.no_grad():
        for data in dataloader:
            inputs, targets = data
            inputs, targets = inputs.to(device), targets.to(device)
            outputs = model(inputs)
            # Implement evaluation metrics here
            pass

# Argument parser
def get_args():
    parser = argparse.ArgumentParser(description="Run NeRF on a dataset")
    parser.add_argument("--iterations", type=int, default=10000, help="Number of iterations for training")
    parser.add_argument("--train", action='store_true', help="Flag to indicate training mode")
    parser.add_argument("--test", action='store_true', help="Flag to indicate test mode")
    parser.add_argument("--output_dir", type=str, default="output", help="Directory to store outputs")
    parser.add_argument("--model_dir", type=str, default="models", help="Directory to store model checkpoints")
    parser.add_argument("--data_path", type=str, required=True, help="Path to the dataset")
    return parser.parse_args()

# Main function
def main():
    args = get_args()

    # Create output and model directories if they don't exist
    os.makedirs(args.output_dir, exist_ok=True)
    os.makedirs(args.model_dir, exist_ok=True)

    # Load data
    dataloader = data_loader(args.data_path)

    # Initialize model
    model = NeRFModel().to(device)

    # Train or test
    if args.train:
        train(model, dataloader, args.iterations, args.model_dir)
    if args.test:
        test(model, dataloader)

if __name__ == "__main__":
    main()
    