requirements.txt
PyOpenGL==3.1.5
PyGLM==2.5.0
numpy==1.23.3
imgui[glfw]
objloader
opencv-python
pywavefront
pyglet
tqdm